change build tags
