+++ channelz/service/service_test.go
@@ -1,4 +1,4 @@
-// +build amd64,linux
+// +build linux
 
 /*
  *
