+++ channelz/util_nonlinux_pre_go19.go
@@ -1,4 +1,4 @@
-// +build !darwin,!dragonfly,!freebsd,!linux,!netbsd,!openbsd,!solaris
+// +build !linux !go1.9
 
 /*
  *
@@ -18,13 +18,9 @@
  *
  */
 
-package service
+package channelz
 
-import (
-	"google.golang.org/grpc/channelz"
-	channelzpb "google.golang.org/grpc/channelz/grpc_channelz_v1"
-)
-
-func sockoptToProto(skopts *channelz.SocketOptionData) []*channelzpb.SocketOption {
+// GetSocketOption gets the socket option info of the conn.
+func GetSocketOption(c interface{}) *SocketOptionData {
 	return nil
 }
